% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterEval-package.R
\docType{package}
\name{ClusterEval}
\alias{ClusterEval}
\title{ClusterEval - Evaluating clusters and clustering}
\description{
A toolbox for exploring and evaluating clustering. Not much here yet.
}
\details{
Home on Github: \url{https://github.com/BioInfoToys/ClusterEval}
}
